# api/internal/swagger/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact: {}
  description: API for managing Allure test reports.
  title: AllureDeck API
  version: 2.0.0
paths:
...
  /projects/{project_id}/known-issues:
    get:
      description: Returns all known issues for a project.
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Filter to active only
        in: query
        name: active_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: List known issues
      tags:
      - known-issues
    post:
      consumes:
      - application/json
      description: Records a new known issue for a project.
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      summary: Create a known issue
      tags:
      - known-issues
  /projects/{project_id}/known-issues/{issue_id}:
    delete:
      description: Permanently removes a known issue record.
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Known Issue ID
        in: path
        name: issue_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Delete a known issue
      tags:
      - known-issues
    put:
      consumes:
      - application/json
      description: Updates ticket_url, description, and/or active status.
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Known Issue ID
        in: path
        name: issue_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Update a known issue
      tags:
      - known-issues
...
  /projects/{project_id}/reports/{report_id}/known-failures:
    get:
      description: Cross-references failed/broken test cases against the known issues
        list.
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Report ID or 'latest'
        in: path
        name: report_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Get known failures for a report
      tags:
      - known-issues
...
swagger: "2.0"